third_party/WebKit/LayoutTests/fast/canvas/canvas-overloads-fillText-expected.txt

Test the behavior of CanvasRenderingContext2D.fillText() when called with different numbers of arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.fillText() threw exception TypeError: Failed to execute 'fillText' on 'CanvasRenderingContext2D': 3 arguments required, but only 0 present..
PASS ctx.fillText('moo') threw exception TypeError: Failed to execute 'fillText' on 'CanvasRenderingContext2D': 3 arguments required, but only 1 present..
PASS ctx.fillText('moo',0) threw exception TypeError: Failed to execute 'fillText' on 'CanvasRenderingContext2D': 3 arguments required, but only 2 present..
PASS ctx.fillText('moo',0,0) is undefined
PASS ctx.fillText('moo',0,0,0) is undefined
PASS ctx.fillText('moo',0,0,0,0) is undefined
PASS successfullyParsed is true

TEST COMPLETE
